made common super class for 3 IT's and fixed checkstyle
